#1382 - moved isEmpty/hasValue logic to DashboardReportFilter that uses specific value types to identify whether value is present or not. Support date range based on Range value type. Separate V1 and V2 code in some other places. from dev settings also clear custom dashboard cache. hash based on filter configs JSON using jshashes library. logger to allow multiple arguments like console log. removed unused functional methods from General and move logging related method be internal to General.

+++ packages/openchs-android/src/service/DashboardCacheService.js
@@ -39,7 +39,8 @@ class DashboardCacheService extends BaseService {
         const db = this.db;
         this.db.write(() => {
             const dashboardCache = this.findOnly();
-            db.delete(dashboardCache);
+            if (!_.isNil(dashboardCache))
+                db.delete(dashboardCache);
         });
     }
 }
